refactor: custom hook, fetch함수 분리하고 적용
