+++ src/api/fetchThemeProduct.ts
@@ -2,8 +2,6 @@ import axios from 'axios';
 import { API_BASE } from '@/constant/constant';
 import type { Product } from '@/types';
 
-import { useInfiniteQuery } from '@tanstack/react-query';
-
 export async function fetchThemeProduct(
   themeId: string,
   cursor?: number,
@@ -26,12 +24,3 @@ interface ThemeProductResponse {
   cursor: number;
   hasMoreList: boolean;
 }
-
-export const useFetchThemeProduct = (themeId: string) => {
-  return useInfiniteQuery({
-    queryKey: ['themeProduct', themeId],
-    queryFn: ({ pageParam = 0 }) => fetchThemeProduct(themeId, pageParam, 10),
-    getNextPageParam: (lastPage) => (lastPage.hasMoreList ? lastPage.cursor : undefined),
-    initialPageParam: 0,
-  });
-};
