+++ src/hooks/useFetchProductSummary.ts
@@ -1,0 +1,9 @@
+import { useQuery } from '@tanstack/react-query';
+import { fetchProductSummary } from '@/api/fetchProductSummary';
+
+export function useFetchProductSummary(productId: string) {
+  return useQuery({
+    queryKey: ['productSummary', productId],
+    queryFn: () => fetchProductSummary(productId),
+  });
+}
