+++ src/hooks/useFetchThemeProduct.ts
@@ -1,0 +1,11 @@
+import { useInfiniteQuery } from '@tanstack/react-query';
+import { fetchThemeProduct } from '@/api/fetchThemeProduct';
+
+export const useFetchThemeProduct = (themeId: string) => {
+  return useInfiniteQuery({
+    queryKey: ['themeProduct', themeId],
+    queryFn: ({ pageParam = 0 }) => fetchThemeProduct(themeId, pageParam, 10),
+    getNextPageParam: (lastPage) => (lastPage.hasMoreList ? lastPage.cursor : undefined),
+    initialPageParam: 0,
+  });
+};
