+++ src/hooks/useOrderMutation.ts
@@ -1,0 +1,9 @@
+import { useMutation } from '@tanstack/react-query';
+import { fetchOrder, type OrderRequest } from '@/api/fetchOrder';
+
+export function useOrderMutation() {
+  return useMutation({
+    mutationFn: ({ orderData, authToken }: { orderData: OrderRequest; authToken: string }) =>
+      fetchOrder(orderData, authToken),
+  });
+}
